requirements.txt
fastapi
google-api-python-client
google-auth
google-auth-oauthlib
langchain_core
langchain_nomic
nltk>=3.8
nomic
openai>=1.0.0
pdf2image
pillow
pytesseract
python-dotenv
pypdf
pyMuPDF
qdrant-client==1.10.1
rank_bm25
requests
scikit-learn
tenacity
tiktoken>=0.5.1
tqdm
uvicorn